revert(disagg): remove experimental decode session pool
