+++ docs_new/docs/advanced_features/pd_disaggregation.mdx
@@ -227,33 +227,14 @@ Please be aware that this setting will cause prefill instances to take a longer 
       <td style={{padding: "9px 12px", backgroundColor: "rgba(255,255,255,0.02)"}}>`300`</td>
     </tr>
     <tr>
-      <td style={{padding: "9px 12px", fontWeight: 500, backgroundColor: "rgba(255,255,255,0.02)"}}>**`SGLANG_DISAGGREGATION_DECODE_SESSION_POOL_SIZE`**</td>
-      <td style={{padding: "9px 12px", backgroundColor: "rgba(255,255,255,0.05)"}}>Experimental Mooncake-only option that creates this many TransferEngine sessions per decode rank. Every session registers the same KV buffers.</td>
-      <td style={{padding: "9px 12px", backgroundColor: "rgba(255,255,255,0.02)"}}>`1`</td>
-    </tr>
-    <tr>
       <td style={{padding: "9px 12px", fontWeight: 500, backgroundColor: "rgba(255,255,255,0.02)"}}>**`SGLANG_DISAGGREGATION_KV_TRANSFER_CHUNK_SIZE`**</td>
       <td style={{padding: "9px 12px", backgroundColor: "rgba(255,255,255,0.05)"}}>Maximum KV indices in one Mooncake transfer work item. Set this for long cached prefixes to avoid one oversized synchronous transfer; <code>0</code> disables splitting.</td>
-      <td style={{padding: "9px 12px", backgroundColor: "rgba(255,255,255,0.02)"}}>`0`</td>
-    </tr>
-    <tr>
-      <td style={{padding: "9px 12px", fontWeight: 500, backgroundColor: "rgba(255,255,255,0.02)"}}>**`SGLANG_DISAGGREGATION_DECODE_ROOMS_PER_SESSION`**</td>
-      <td style={{padding: "9px 12px", backgroundColor: "rgba(255,255,255,0.05)"}}>Soft active-room target for least-loaded session assignment. When every session reaches the target, requests continue on the least-loaded session. Set to <code>0</code> to disable the target.</td>
       <td style={{padding: "9px 12px", backgroundColor: "rgba(255,255,255,0.02)"}}>`0`</td>
     </tr>
   </tbody>
 </table>
 
 If a greater mean TTFT is acceptable, you can `export SGLANG_DISAGGREGATION_WAITING_TIMEOUT=600` (10 minutes) to relax the timeout condition.
-
-For example, to spread 256 active rooms across four Mooncake sessions on each decode rank, set:
-
-```bash
-export SGLANG_DISAGGREGATION_DECODE_SESSION_POOL_SIZE=4
-export SGLANG_DISAGGREGATION_DECODE_ROOMS_PER_SESSION=64
-export SGLANG_DISAGGREGATION_KV_TRANSFER_CHUNK_SIZE=8192
-```
-
 
 ## Heterogeneous TP with GPU Staging Buffer
