+++ python/sglang/srt/disaggregation/mooncake/conn.py
@@ -95,70 +95,6 @@ def _parse_transfer_status_message(
     return bootstrap_room, status, prefill_rank
 
 
-class MooncakeDecodeSessionPool:
-    """Least-loaded assignment of decode rooms to Mooncake sessions.
-
-    ``rooms_per_session`` is deliberately a soft target. Rejecting or delaying
-    a request here would require cross-rank admission coordination; for this
-    experimental pool, overflow keeps serving on the least-loaded session.
-    """
-
-    def __init__(self, engines: list, rooms_per_session: int = 0):
-        if not engines:
-            raise ValueError(
-                "Mooncake decode session pool requires at least one engine"
-            )
-        if rooms_per_session < 0:
-            raise ValueError("rooms_per_session must be non-negative")
-        self.engines = engines
-        self.rooms_per_session = rooms_per_session
-        self._active_rooms = [0] * len(engines)
-        self._room_to_session: dict[int, int] = {}
-        self._lock = threading.Lock()
-        self._overflow_warned = False
-
-    def acquire(self, room: int) -> Tuple[int, str]:
-        with self._lock:
-            existing = self._room_to_session.get(room)
-            if existing is not None:
-                return existing, self.engines[existing].get_session_id()
-
-            candidates = range(len(self.engines))
-            if self.rooms_per_session > 0:
-                below_target = [
-                    i
-                    for i, count in enumerate(self._active_rooms)
-                    if count < self.rooms_per_session
-                ]
-                if below_target:
-                    candidates = below_target
-                elif not self._overflow_warned:
-                    logger.warning(
-                        "All %d Mooncake decode sessions reached the soft target "
-                        "of %d active rooms; continuing on the least-loaded session",
-                        len(self.engines),
-                        self.rooms_per_session,
-                    )
-                    self._overflow_warned = True
-
-            session_index = min(candidates, key=self._active_rooms.__getitem__)
-            self._room_to_session[room] = session_index
-            self._active_rooms[session_index] += 1
-            return session_index, self.engines[session_index].get_session_id()
-
-    def release(self, room: int) -> None:
-        with self._lock:
-            session_index = self._room_to_session.pop(room, None)
-            if session_index is None:
-                return
-            self._active_rooms[session_index] -= 1
-            assert self._active_rooms[session_index] >= 0
-
-    def active_room_counts(self) -> Tuple[int, ...]:
-        with self._lock:
-            return tuple(self._active_rooms)
-
-
 # decode
 @dataclasses.dataclass
 class TransferInfo:
@@ -368,111 +304,40 @@ class MooncakeKVManager(CommonKVManager):
 
     def init_engine(self):
         self.engine = get_mooncake_transfer_engine()
-        self.engines = [self.engine]
-
-        pool_size = 1
-        rooms_per_session = 0
-        if self.disaggregation_mode == DisaggregationMode.DECODE:
-            pool_size = envs.SGLANG_DISAGGREGATION_DECODE_SESSION_POOL_SIZE.get()
-            rooms_per_session = (
-                envs.SGLANG_DISAGGREGATION_DECODE_ROOMS_PER_SESSION.get()
-            )
-        if pool_size < 1:
-            raise ValueError(
-                "SGLANG_DISAGGREGATION_DECODE_SESSION_POOL_SIZE must be >= 1"
-            )
-        if rooms_per_session < 0:
-            raise ValueError(
-                "SGLANG_DISAGGREGATION_DECODE_ROOMS_PER_SESSION must be >= 0"
-            )
-
-        if pool_size > 1:
-            from sglang.srt.distributed.device_communicators.mooncake_transfer_engine import (
-                MooncakeTransferEngine,
-            )
-
-            logger.warning(
-                "Enabling experimental Mooncake decode session pool: "
-                "sessions=%d rooms_per_session=%d",
-                pool_size,
-                rooms_per_session,
-            )
-            for _ in range(1, pool_size):
-                self.engines.append(
-                    MooncakeTransferEngine(
-                        hostname=self.engine.hostname,
-                        gpu_id=self.engine.gpu_id,
-                        # The shared engine has already resolved a per-GPU HCA
-                        # string, which can be reused directly by each clone.
-                        ib_device=self.engine.ib_device,
-                    )
-                )
-
-        self.decode_session_pool = MooncakeDecodeSessionPool(
-            self.engines, rooms_per_session=rooms_per_session
-        )
-        self._session_registration_lock = threading.Lock()
-        self._session_route_registrations: dict[int, tuple[dict, set[str]]] = {}
-        if pool_size > 1:
-            logger.info(
-                "Mooncake decode session pool endpoints: %s",
-                [engine.get_session_id() for engine in self.engines],
-            )
-
-    def _batch_register_all_engines(self, ptrs: List[int], lens: List[int]) -> int:
-        for engine in self.engines:
-            ret = engine.batch_register(ptrs, lens)
-            if ret not in (None, 0):
-                return ret
-        return 0
-
-    def _batch_deregister_all_engines(self, ptrs: List[int]) -> None:
-        for engine in self.engines:
-            engine.batch_deregister(ptrs)
 
     def register_buffer_to_engine(self):
         # Batch register KV data buffers
         if self.kv_args.kv_data_ptrs and self.kv_args.kv_data_lens:
-            ret = self._batch_register_all_engines(
+            self.engine.batch_register(
                 self.kv_args.kv_data_ptrs, self.kv_args.kv_data_lens
             )
-            if ret != 0 and len(self.engines) > 1:
-                raise RuntimeError("Failed to register KV buffers to session pool")
 
         # Batch register auxiliary data buffers
         if self.kv_args.aux_data_ptrs and self.kv_args.aux_data_lens:
-            ret = self._batch_register_all_engines(
+            self.engine.batch_register(
                 self.kv_args.aux_data_ptrs, self.kv_args.aux_data_lens
             )
-            if ret != 0 and len(self.engines) > 1:
-                raise RuntimeError("Failed to register aux buffers to session pool")
 
         for ptrs, lens in zip(
             self.kv_args.state_data_ptrs, self.kv_args.state_data_lens
         ):
             if ptrs and lens:
-                ret = self._batch_register_all_engines(ptrs, lens)
-                if ret != 0 and len(self.engines) > 1:
-                    raise RuntimeError(
-                        "Failed to register state buffers to session pool"
-                    )
+                self.engine.batch_register(ptrs, lens)
 
     def deregister_buffer_to_engine(self):
         if self.kv_args.kv_data_ptrs:
-            self._batch_deregister_all_engines(self.kv_args.kv_data_ptrs)
+            self.engine.batch_deregister(self.kv_args.kv_data_ptrs)
 
         if self.kv_args.aux_data_ptrs:
-            self._batch_deregister_all_engines(self.kv_args.aux_data_ptrs)
+            self.engine.batch_deregister(self.kv_args.aux_data_ptrs)
 
         for ptrs in self.kv_args.state_data_ptrs or []:
             if ptrs:
-                self._batch_deregister_all_engines(ptrs)
+                self.engine.batch_deregister(ptrs)
 
         if hasattr(self, "connection_pool"):
             with self.connection_lock:
                 self.connection_pool.clear()
-        with self._session_registration_lock:
-            self._session_route_registrations.clear()
 
     # ------------------------------------------------------------------
     # Staging buffer methods (all delegate to staging_handler.py)
@@ -495,7 +360,7 @@ class MooncakeKVManager(CommonKVManager):
         )
 
         self._staging_ctx.buffers = init_staging_buffers(
-            lambda ptr, size: self._batch_register_all_engines([ptr], [size]),
+            lambda ptr, size: self.engine.batch_register([ptr], [size]),
             self.kv_args,
             count,
         )
@@ -507,7 +372,7 @@ class MooncakeKVManager(CommonKVManager):
         )
 
         self._staging_ctx.allocator = init_staging_allocator(
-            lambda ptr, size: self._batch_register_all_engines([ptr], [size]),
+            lambda ptr, size: self.engine.batch_register([ptr], [size]),
             self.kv_args,
         )
         self.kv_buffer_tensors = None
@@ -2931,44 +2796,6 @@ class MooncakeKVManager(CommonKVManager):
     def get_session_id(self):
         return self.engine.get_session_id()
 
-    def acquire_decode_session(self, bootstrap_room: int) -> Tuple[int, str]:
-        assert self.disaggregation_mode == DisaggregationMode.DECODE
-        return self.decode_session_pool.acquire(bootstrap_room)
-
-    def release_decode_session(self, bootstrap_room: int) -> None:
-        if self.disaggregation_mode == DisaggregationMode.DECODE:
-            self.decode_session_pool.release(bootstrap_room)
-
-    def claim_session_route_registration(
-        self, bootstrap_info: dict, session_id: str
-    ) -> bool:
-        """Claim one session/route registration, returning False if known.
-
-        Registrations are keyed by the identity of the cached bootstrap-info
-        object without mutating it. If heartbeat handling drops and later
-        recreates a route, the new object has a new identity and every session
-        is registered again with the restarted prefill process.
-        """
-        with self._session_registration_lock:
-            route_id = id(bootstrap_info)
-            entry = self._session_route_registrations.get(route_id)
-            if entry is None or entry[0] is not bootstrap_info:
-                entry = (bootstrap_info, set())
-                self._session_route_registrations[route_id] = entry
-            registered = entry[1]
-            if session_id in registered:
-                return False
-            registered.add(session_id)
-            return True
-
-    def unclaim_session_route_registration(
-        self, bootstrap_info: dict, session_id: str
-    ) -> None:
-        with self._session_registration_lock:
-            entry = self._session_route_registrations.get(id(bootstrap_info))
-            if entry is not None and entry[0] is bootstrap_info:
-                entry[1].discard(session_id)
-
     def _on_heartbeat_success(self, bootstrap_addr: str):
         current_rooms = self.addr_to_rooms_tracker[bootstrap_addr].copy()
         for bootstrap_room in current_rooms:
@@ -3155,30 +2982,12 @@ class MooncakeKVReceiver(CommonKVReceiver):
         bootstrap_addr: str,
         bootstrap_room: Optional[int] = None,
     ):
-        self.session_index, self.session_id = mgr.acquire_decode_session(bootstrap_room)
-        self._session_released = False
+        self.session_id = mgr.get_session_id()
         self.init_time = None
-        try:
-            super().__init__(mgr, bootstrap_addr, bootstrap_room)
-        except Exception:
-            self._release_session()
-            raise
-
-    def _setup_bootstrap_infos(self):
-        super()._setup_bootstrap_infos()
-        # CommonKVReceiver registers only when a route is first cached. A
-        # later room may select another session while reusing that route, so
-        # ensure the selected session is registered as well. Per-route markers
-        # inside _register_kv_args make this idempotent.
-        if self.bootstrap_infos is not None and len(self.kv_mgr.engines) > 1:
-            self._register_kv_args()
+        super().__init__(mgr, bootstrap_addr, bootstrap_room)
 
     def _register_kv_args(self):
         for bootstrap_info in self.bootstrap_infos:
-            if not self.kv_mgr.claim_session_route_registration(
-                bootstrap_info, self.session_id
-            ):
-                continue
             packed_kv_data_ptrs = b"".join(
                 struct.pack("Q", ptr) for ptr in self.kv_mgr.kv_args.kv_data_ptrs
             )
@@ -3214,35 +3023,29 @@ class MooncakeKVReceiver(CommonKVReceiver):
                 packed_staging_base_ptr = b""
                 staging_total_size_str = b""
 
-            try:
-                sock, lock = self._connect_to_bootstrap_server(bootstrap_info)
-                with lock:
-                    sock.send_multipart(
-                        [
-                            "None".encode("ascii"),
-                            self.kv_mgr.local_ip.encode("ascii"),
-                            str(self.kv_mgr.rank_port).encode("ascii"),
-                            self.session_id.encode("ascii"),
-                            packed_kv_data_ptrs,
-                            packed_aux_data_ptrs,
-                            packed_state_data_ptrs,
-                            dst_tp_rank,
-                            dst_attn_tp_size,
-                            dst_kv_item_len,
-                            packed_state_item_lens,
-                            packed_state_dim_per_tensor,
-                            enable_hisparse,
-                            dst_dcp_size,
-                            dst_dcp_rank,
-                            packed_staging_base_ptr,
-                            staging_total_size_str,
-                        ]
-                    )
-            except Exception:
-                self.kv_mgr.unclaim_session_route_registration(
-                    bootstrap_info, self.session_id
+            sock, lock = self._connect_to_bootstrap_server(bootstrap_info)
+            with lock:
+                sock.send_multipart(
+                    [
+                        "None".encode("ascii"),
+                        self.kv_mgr.local_ip.encode("ascii"),
+                        str(self.kv_mgr.rank_port).encode("ascii"),
+                        self.session_id.encode("ascii"),
+                        packed_kv_data_ptrs,
+                        packed_aux_data_ptrs,
+                        packed_state_data_ptrs,
+                        dst_tp_rank,
+                        dst_attn_tp_size,
+                        dst_kv_item_len,
+                        packed_state_item_lens,
+                        packed_state_dim_per_tensor,
+                        enable_hisparse,
+                        dst_dcp_size,
+                        dst_dcp_rank,
+                        packed_staging_base_ptr,
+                        staging_total_size_str,
+                    ]
                 )
-                raise
 
     def send_metadata(
         self,
@@ -3305,24 +3108,6 @@ class MooncakeKVReceiver(CommonKVReceiver):
                 return timeout_result
 
         return status
-
-    def _release_session(self) -> None:
-        if self._session_released:
-            return
-        self.kv_mgr.release_decode_session(self.bootstrap_room)
-        self._session_released = True
-
-    def clear(self) -> None:
-        try:
-            super().clear()
-        finally:
-            self._release_session()
-
-    def abort(self):
-        try:
-            super().abort()
-        finally:
-            self._release_session()
 
     def failure_exception(self):
         if self.conclude_state is None:
